Improvements and progress, +1 button and model progress
App now stores internal file with the date of the latest json obtained from lichess. If the current date is equal to the date of the latest json "pull" made to the URL, it will not bother making the request/connection.

Made a button to launch new PuzzleSolvingActivity, the intent() and the putExtra() with the puzzle and solution for it.

Added more strings to TranslationEditor and incorporated them into the toasts. Made overloaded logs and toasts for Array<String>

For the Model, I started a sketch of the logic and handling of the pieces moves in ChessPieces.kt

Now the initial start of a game is in Board.kt in the form of an array of ChessPieces. This array is then accessed in read only by BoardView and interprets its char values to convert it into a R.drawable.ic[color]_[pieceName]. Correctly dividing Model from View, which wasn't happening before when priting the starting point of the chess game