Fixed 2 bugs regarding MainActivity and clicking progress for PuzzleSolvingActivity
Fixed bug where it wouldn't get today's date if our internal txt file already has the date. If the app was closed and is open from the beggining from onCreate, the data for the game wouldn't be in memory even though our internal file has the date of today. And thus, our method getTodaysGame() woulnd't let us get the data. So. I changed the first if() in this function, which will deny the request only if our data objects aren't null (and therefore, are in memory)

Fixed our snackBar method call and definition method to snackBar(R.string.jsonError, ::getTodaysGame) and snackBar(stringID: Int, function: KFunction0<Unit>) respectively

Added BoardView to activity_puzzle_solving.xml and added a test-sketch Listener 4 each tile on PuzzleSolvingActivity. It will write in log the color of the tile that was clicked.

Simplified the function getPiece(piece: Piece) to cut number of lines, but increase line lenght in BoardView.kt