Can move any piece to any valid location + little stuff
Fixed error for that in the first installement of this app in a device, the DATEFILE (.txt) would not be found and it would throw FileNotFoundException. So I added a try-catch. Catch contain " this.openFileOutput(DATEFILE, MODE_PRIVATE)" which creates a new file and calls the function again. Started making methods to interpret and load the puzzle into the board