Can now interpret and move other pieced loaded from json
Except for 0-0, 0-0-0 and movements with x on its encoding which kills pieces. Also Im repeating code on this interpretMove method. This will be handled near the end. For now, Im trying to organize and check for every case, and only then I will simplify. This is because by then, I will be able to analyse patterns in order to simplify my code.