Simplified castle move logic and resolved data consistency (readme)
The great interpretMove method in Board.kt is in its final stages of completeness. The handling of "O-O-O" and "O-O" was simplified by formulating the indexing differences between the 2. This is explained in a comment at line 212, which I will delete after this commit to save me some space.

I noticed that when I entered the 2nd activity, and pressed the back button and hit continue button, the loadGame() in PuzzleSolvingActivity would fail. And it would actually save any movements I had done before clicking the back button. So, when I clicked continue, it was like the state of the pieces was saved. I'm not sure if I recall it correctly, but according to my analysis, it turns out that the Position objects were somehow kept in memory for each piece in the array. It turns out it was related to my variable chessPiecesTablePositions in Board.kt which was defined outside the class. Which seems to be a static variable to me, and thus this problem

So I passed my var array called chessPiecesTablePositions to inside the class which is a copy of the VAL array called startingChessPiecesTablePositions. And I adapted the companion object to this alteration.

With this, I hope that the data in my array of Pieces is consistent. In resume, the user can go back to first activity, and click continue to the next, and the puzzle will reloaded from the start successfully.