Added ViewModel for PuzzleSolvingActivity (read description)
The PuzzleSolvingAcitivityViewModel stores in memory the variables:
1 - isGameLoaded: MutableLiveData<Boolean>
2 - board: Board
3 - isWhitesPlaying: Boolean

For each variable:
1 - Makes sures that the puzzle in the Array<String>isn't interpreted again
2 - Makes sure the game/board state/positions are maintained when it goes through state and configuration changes throughout the activity lifecycle
3 - Keeps consistent the rule of alternating plays between the white and black pieces is maintained

Also changed the setOnClickListener for each Tile so that the movements are more properly validated according to the selected piece rules and the current color playing. And added some logs() to it.

One of the remaining things to do is to keep the states of some pieces consistent to the moves interpreted in the puzzle (per example, the pawn can only move 2 positions only in its first move) and some other chess rules like only the horse can jump over other pieces

Also removed the Boolean return value in getTodaysGame and fixed the snackBar so that it's run on isGameReady.observe when it returns false. I realized that methods in ViewMethods shouldnt return values (since that it doesnt work quite well with I/O) and I should instead use observers (using LiveData). The snackBar will now appear succesfuly if there is no wifi or mobile data turned on.