Added support for move encoded with line number (puzzle id = CiAThnSX)
In this puzzle, there was a conflict with 2 knights in the same column. I documented this in Notes.png. The move encoding is N6h5. Which my method wasn't ready to do since I never crossed with this case, but now I changed interpretMove method so that it uses the line indication (6) to filter the right knight (or any other piece) to move to. So now it loaded the game successfully